chore: isolate Next app-render compatibility
